move use/audio.ts
