+++ src/components/UI/ChromeAudio.vue
@@ -37,7 +37,7 @@
 <script lang="ts">
 import { defineComponent } from 'vue'
 import useFileMeta from '@/use/fileMeta'
-import useAudio from './use/audio'
+import useAudio from '@/use/audio'
 import AudioPlayerPlayButton from './AudioPlayer/AudioPlayerPlayButton.vue'
 import AudioPlayerTime from './AudioPlayer/AudioPlayerTime.vue'
 import AudioPlayerTimeSlider from './AudioPlayer/AudioPlayerTimeSlider.vue'
